SLBR_PR/requirements.txt
albumentations==0.4.5
certifi==2021.10.8
charset-normalizer==2.0.12
cloudpickle==2.0.0
cycler==0.11.0
idna==3.3
imageio==2.19.2
imgaug==0.2.6
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==2.2.2
networkx==2.6.3
numpy==1.19.1
opencv-python==4.5.5.64
Pillow==9.1.1
progress==1.5
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.17.2
scikit-learn==1.0
scipy==1.4.1
six==1.16.0
sklearn==0.0
tensorboardX==2.4
threadpoolctl==3.1.0
tifffile==2021.11.2
tqdm==4.46.1
typing_extensions==4.2.0
urllib3==1.26.9
wincertstore==0.2
